releasetools: Fix an issue where copytree appends "/root" to some directories
error:
Traceback (most recent call last):
  File "build_image.py", line 949, in <module>
  File "build_image.py", line 941, in main
  File "build_image.py", line 506, in BuildImage
  File "build_image.py", line 175, in SetUpInDirAndFsConfig
  File "internal/stdlib/shutil.py", line 554, in copytree
FileNotFoundError: [Errno 2] No such file or directory: 'out/target/product/xx/root'

+++ tools/releasetools/build_image.py
@@ -170,12 +170,16 @@ def SetUpInDirAndFsConfig(origin_in, prop_dict):
   # Construct a staging directory of the root file system.
   in_dir = common.MakeTempDir()
   root_dir = prop_dict.get("root_dir")
-  if root_dir:
-    shutil.rmtree(in_dir)
-    shutil.copytree(root_dir, in_dir, symlinks=True)
+
+  if root_dir and os.path.exists(root_dir):
+      shutil.rmtree(in_dir)
+      shutil.copytree(root_dir, in_dir, symlinks=True)
   in_dir_system = os.path.join(in_dir, "system")
-  shutil.rmtree(in_dir_system, ignore_errors=True)
-  shutil.copytree(origin_in, in_dir_system, symlinks=True)
+  in_dir_system = in_dir_system.rstrip(os.path.sep + "root")
+
+  if os.path.exists(in_dir_system):
+      shutil.rmtree(in_dir_system, ignore_errors=True)
+      shutil.copytree(origin_in, in_dir_system, symlinks=True)
 
   # Change the mount point to "/".
   prop_dict["mount_point"] = "/"
